fix: Adapt CSV processing for multiple containers
The `ews process-csv` command processes the CSV to generate an image
mirror mapping file. In order to do this it finds image references in
relevant fields, including the containers field of the operator
Deployment. In the past, operators would have only one container (the
controller manager itself), but newer operators, since the release of
operator-sdk 1.0, include a proxy container.

The CSV processing logic would take only the first container for
processing, thi change iterates through the containers list and
processes them all.

Adapt the unit tests to consider this situation